requirements.txt
asgiref==3.7.2
Django==4.2.6
django-enumfields==2.1.1
djangorestframework==3.14.0
networkx==3.2
psycopg2==2.9.9
pytz==2023.3.post1
sqlparse==0.4.4
typing_extensions==4.8.0
